format: 마이 페이지 유저 정보 컴포넌트 코드 구조 변경
