+++ src/components/MyPage/UserInfo/index.tsx
@@ -1,43 +1,38 @@
-import { ReactElement, ReactNode } from 'react';
 import ProfileImage from 'components/common/ProfileImage';
-import TPlatform from 'types/platform';
+import PATH from 'constants/path';
+import IUserData from 'pages/MyPage/types';
 import * as S from './style';
 
-interface IUserActivity {
-  children: string;
-  count: number;
-  to: string;
-}
-
-function UserInfo({ children }: { children: ReactNode }) {
-  return <S.Wrapper>{children}</S.Wrapper>;
-}
-
-function UserName({ children }: { children: string }) {
-  return <S.Name>{children}</S.Name>;
-}
-
-function LoginInfo({ children, platform }: { children: string; platform: TPlatform }) {
-  return <S.LoginInfo platform={platform}>{children}</S.LoginInfo>;
-}
-
-function UserActivities({ children }: { children: ReactElement<IUserActivity>[] }) {
-  return <S.ActivitiesWrapper>{children}</S.ActivitiesWrapper>;
-}
-
-function UserActivity({ children, count, to }: IUserActivity) {
+function UserInfoSection({
+  profileImageUrl,
+  nickname,
+  platform,
+  email,
+  postedTreesCount,
+  savedTreesCount,
+  reviewsCount,
+}: IUserData) {
   return (
-    <S.Activity to={to}>
-      <p>{count}</p>
-      <p>{children}</p>
-    </S.Activity>
+    <S.Wrapper>
+      <ProfileImage src={profileImageUrl} />
+      <S.Name>{nickname}</S.Name>
+      <S.LoginInfo $platform={platform}>{email}</S.LoginInfo>
+      <S.ActivitiesWrapper>
+        <S.Activity to={`.${PATH.registInfoPage}`}>
+          <p>{postedTreesCount}</p>
+          <p>등록한 트리</p>
+        </S.Activity>
+        <S.Activity to={`${PATH.saveTreePage}`}>
+          <p>{savedTreesCount}</p>
+          <p> 저장한 트리</p>
+        </S.Activity>
+        <S.Activity to={`${PATH.reviewPage}`}>
+          <p>{reviewsCount}</p>
+          <p>내가 쓴 후기</p>
+        </S.Activity>
+      </S.ActivitiesWrapper>
+    </S.Wrapper>
   );
 }
 
-UserInfo.Profile = ProfileImage;
-UserInfo.Name = UserName;
-UserInfo.Login = LoginInfo;
-UserInfo.Activities = UserActivities;
-UserInfo.Activity = UserActivity;
-
-export default UserInfo;
+export default UserInfoSection;
